# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bill's Pro API",
        "description": "API documentation for Bill's Pro - Crypto and Bill Payment System",
        "contact": {
            "name": "API Support",
            "email": "support@billspro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/virtual-cards": {
            "get": {
                "tags": [
                    "Virtual Cards"
                ],
                "summary": "Get user's virtual cards",
                "description": "Get all active virtual cards for the authenticated user.",
                "operationId": "99ca84c76187e54ba3224e8add2f49e0",
                "responses": {
                    "200": {
                        "description": "Cards retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Virtual Cards"
                ],
                "summary": "Create virtual card",
                "description": "Create a new virtual card. Requires card creation fee of $3.",
                "operationId": "3c5278ce0513afc80c49f302df082d37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_name",
                                    "payment_wallet_type"
                                ],
                                "properties": {
                                    "card_name": {
                                        "type": "string",
                                        "example": "Qamardeen Abdulmalik"
                                    },
                                    "card_color": {
                                        "type": "string",
                                        "enum": [
                                            "green",
                                            "brown",
                                            "purple"
                                        ],
                                        "example": "green",
                                        "nullable": true
                                    },
                                    "payment_wallet_type": {
                                        "type": "string",
                                        "enum": [
                                            "naira_wallet",
                                            "crypto_wallet"
                                        ],
                                        "example": "naira_wallet"
                                    },
                                    "payment_wallet_currency": {
                                        "type": "string",
                                        "example": "NGN",
                                        "nullable": true
                                    },
                                    "billing_address_street": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "billing_address_city": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "billing_address_state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "billing_address_country": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "billing_address_postal_code": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Card created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient balance or invalid request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/virtual-cards/{id}": {
            "get": {
                "tags": [
                    "Virtual Cards"
                ],
                "summary": "Get card details",
                "description": "Get detailed information about a specific virtual card including card number, CVV, and expiry.",
                "operationId": "62b7a35f3044a434e872c5281c9abe72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}